// src/typings/nodes.ts
export enum FIGMA_NODE {
  BOOLEAN_OPERATION = 'BOOLEAN_OPERATION',
  CODE_BLOCK = 'CODE_BLOCK',
  COMPONENT = 'COMPONENT',
  COMPONENT_SET = 'COMPONENT_SET',
  CONNECTOR = 'CONNECTOR',
  DOCUMENT = 'DOCUMENT',
  ELLIPSE = 'ELLIPSE',
  EMBED = 'EMBED',
  FRAME = 'FRAME',
  GROUP = 'GROUP',
  HIGHLIGHT = 'HIGHLIGHT',
  INSTANCE = 'INSTANCE',
  LINE = 'LINE',
  LINK_UNFURL = 'LINK_UNFURL',
  MEDIA = 'MEDIA',
  PAGE = 'PAGE',
  POLYGON = 'POLYGON',
  RECTANGLE = 'RECTANGLE',
  SECTION = 'SECTION',
  SHAPE_WITH_TEXT = 'SHAPE_WITH_TEXT',
  SLICE = 'SLICE',
  STAMP = 'STAMP',
  STAR = 'STAR',
  STICKY = 'STICKY',
  TEXT = 'TEXT',
  VECTOR = 'VECTOR',
  WASHI_TAPE = 'WASHI_TAPE',
  WIDGET = 'WIDGET',
  TABLE = 'TABLE',
  TABLE_CELL = 'TABLE_CELL',
}